requirements.txt
flask
tensorflow-cpu
numpy
pandas
gunicorn
requests
